Guarantee Relations Update

+++ src/main/java/com/project/JewelryMS/service/PerformanceService.java
@@ -29,7 +29,7 @@ public class PerformanceService {
 
     public PerformanceResponse createPerformanceReport(CreatePerformanceRequest createPerformanceRequest) {
         //the staffAccountRepository could have a JpaRepository of <Integer>
-        Optional<StaffAccount> staffOptional = staffAccountRepository.findById((long)createPerformanceRequest.getStaffID());
+        Optional<StaffAccount> staffOptional = staffAccountRepository.findById(createPerformanceRequest.getStaffID());
         if(staffOptional.isPresent()) {
             StaffAccount account = staffOptional.get();
             if (account != null) {//This if condition needs revision, should use to connect with Shift to accurately create a performance report
